KFSPTS-21006 More PR feedback changes
