--- src/main/java/edu/cornell/kfs/kew/mail/service/CuEmailContentService.java
@@ -1,9 +1,0 @@
-package edu.cornell.kfs.kew.mail.service;
-
-import org.kuali.kfs.kew.mail.service.EmailContentService;
-
-public interface CuEmailContentService extends EmailContentService {
-
-    boolean isProductionWorkflowEmailModeEnabled();
-
-}
