+++ src/main/resources/edu/cornell/kfs/kew/config/cu-spring-kew.xml
@@ -155,18 +155,22 @@
           lazy-init="true" p:routeHeaderDAO-ref="enDocumentRouteHeaderDAO"
           p:searchableAttributeDAO-ref="enSearchableAttributeDAO"/>
 
-    <!-- Override action list email services to support forcing PROD-style emails in non-PROD environments. -->
-
-    <bean id="enEmailContentService" class="edu.cornell.kfs.kew.mail.service.impl.CuEmailContentServiceImpl"
+    <!--
+        Override of email content service to configure the "deploymentEnvironment" property correctly.
+        Once KualiCo fixes that issue, this email content service override can be removed.
+     -->
+    <bean id="enEmailContentService" class="org.kuali.kfs.kew.mail.service.impl.EmailContentServiceImpl"
           lazy-init="true"
           p:emailService-ref="emailService"
           p:routeHeaderService-ref="enDocumentRouteHeaderService"
-          p:deploymentEnvironment="${environment}"
-          p:parameterService-ref="parameterService"/>
+          p:deploymentEnvironment="${environment}"/>
+
+    <!-- Override action list email service to support forcing PROD-style emails in non-PROD environments. -->
 
     <bean id="enActionListEmailService"
           class="edu.cornell.kfs.kew.mail.service.impl.CuActionListEmailServiceImpl"
           p:deploymentEnvironment="${environment}" p:emailContentGenerator-ref="enEmailContentService"
-          p:emailService-ref="emailService" p:scheduler-ref="scheduler"/>
+          p:emailService-ref="emailService" p:scheduler-ref="scheduler"
+          p:parameterService-ref="parameterService"/>
 
 </beans>
